requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
attrs==25.3.0
ccxt==4.4.71
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
frozenlist==1.5.0
idna==3.10
logzero==1.7.0
multidict==6.3.0
pandas==2.2.2
propcache==0.3.1
pycares==4.5.0
pycparser==2.22
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
setuptools==78.1.0
six==1.17.0
smartapi-python @ git+https://github.com/angel-one/smartapi-python.git@a91d3f0e1553151865dc790e7f67fa44a7ea4214
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
websocket-client==1.8.0
yarl==1.18.3
